pyproject.toml
[project]
name = "route-watch"
version = "0.1.0"
description = "CLI tool for monitoring route traffic congestion"
authors = [
    {name = "route_watch"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[project.scripts]
route_watch = "route_watch.cli:cli"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
mypy = "^1.17.1"
ruff = "^0.12.12"
types-pyyaml = "^6.0.12.20250822"
types-toml = "^0.10.8.20240310"

